fix compile warning in attr_node_func.c on x86_64

+++ src/lib/Libattr/attr_node_func.c
@@ -1152,7 +1152,7 @@ int set_note_str(
   if (nsize > MAX_NOTE)
     {
     sprintf(log_buffer,"Warning: Client attempted to set note with len (%d) > MAX_NOTE (%d)",
-      nsize,
+      (int)nsize,
       MAX_NOTE);
 
     log_record(
